refactor(serverpod_shared): hoist isProcessAlive to a shared FFI util
Replaces two near-identical shell-out implementations of the same liveness check.
Both now go through FFI, either libc kill(pid, 0) or win32 OpenProcess.
Renamed from processPidIsLive to isProcessAlive.
