+++ packages/serverpod_embedded_postgres/pubspec.yaml
@@ -21,7 +21,6 @@ dependencies:
   postgres: ^3.4.0
   pub_semver: ^2.1.4
   serverpod_shared: 3.5.0-beta.7
-  win32: ^5.13.0
 
 dev_dependencies:
   serverpod_lints: 3.5.0-beta.7
