+++ packages/serverpod_shared/lib/serverpod_shared.dart
@@ -11,5 +11,6 @@ export 'src/migration_exceptions.dart';
 export 'src/password_manager.dart';
 export 'src/util.dart';
 export 'src/utils/file_ex.dart';
+export 'src/utils/process_alive.dart';
 export 'src/utils/secure_directory.dart';
 export 'src/utils/unix_socket.dart';
