+++ templates/pubspecs/packages/serverpod_embedded_postgres/pubspec.yaml
@@ -20,7 +20,6 @@ dependencies:
   postgres: ^3.4.0
   pub_semver: ^2.1.4
   serverpod_shared: SERVERPOD_VERSION
-  win32: ^5.13.0
 
 dev_dependencies:
   serverpod_lints: SERVERPOD_VERSION
